:zap: Improve strat screen
Improve start progress
